get_reference_housenumbers: extract simplify, used in suspicious_streets
